+++ Content/Enemies/Bosses/StoneTitan.cs
@@ -1,4 +1,7 @@
-﻿namespace WellRoundedBalance.Enemies.Bosses
+﻿using EntityStates;
+using RoR2.Skills;
+
+namespace WellRoundedBalance.Enemies.Bosses
 {
     internal class StoneTitan : EnemyBase<StoneTitan>
     {
@@ -108,6 +111,9 @@
 
         private void Changes()
         {
+            ContentAddition.AddEntityState(typeof(ChargeLaserSpam), out _);
+            ContentAddition.AddEntityState(typeof(FireLaserSpam), out _);
+
             var rockController = Utils.Paths.GameObject.TitanRockController.Load<GameObject>();
             var destroyOnTimer = rockController.GetComponent<DestroyOnTimer>();
             destroyOnTimer.duration = 20f;
@@ -121,6 +127,266 @@
             var titanBody = Utils.Paths.GameObject.TitanBody13.Load<GameObject>().GetComponent<CharacterBody>();
             titanBody.baseDamage = 25f;
             titanBody.levelDamage = 5f;
+
+            var laserSD = Utils.Paths.SkillDef.TitanBodyLaser.Load<SkillDef>();
+            laserSD.activationState = new SerializableEntityStateType(typeof(ChargeLaserSpam));
+        }
+    }
+
+    public class ChargeLaserSpam : BaseState
+    {
+        public static float baseDuration = 3f;
+
+        public static float laserMaxWidth = 0.2f;
+
+        public static GameObject effectPrefab = Utils.Paths.GameObject.ChargeGolem.Load<GameObject>();
+
+        public static GameObject laserPrefab = Utils.Paths.GameObject.LaserGolem.Load<GameObject>();
+
+        public static string attackSoundString = "Play_titanboss_R_laser_preshoot";
+
+        public float duration;
+
+        public uint chargePlayID;
+
+        public GameObject chargeEffect;
+
+        public GameObject laserEffect;
+
+        public LineRenderer laserLineComponent;
+
+        public Vector3 laserDirection;
+
+        public float flashTimer = 0f;
+
+        public bool laserOn = true;
+
+        public override void OnEnter()
+        {
+            base.OnEnter();
+            duration = baseDuration / attackSpeedStat;
+            var modelTransform = GetModelTransform();
+            chargePlayID = Util.PlayAttackSpeedSound(attackSoundString, gameObject, attackSpeedStat);
+            if (modelTransform)
+            {
+                var childLocator = modelTransform.GetComponent<ChildLocator>();
+                if (childLocator)
+                {
+                    var laserMuzzle = childLocator.FindChild("MuzzleLaser");
+                    if (laserMuzzle)
+                    {
+                        if (effectPrefab)
+                        {
+                            chargeEffect = Object.Instantiate(effectPrefab, laserMuzzle.position, laserMuzzle.rotation);
+                            chargeEffect.transform.parent = laserMuzzle;
+                            var scaleParticleSystemDuration = chargeEffect.GetComponent<ScaleParticleSystemDuration>();
+                            if (scaleParticleSystemDuration)
+                            {
+                                scaleParticleSystemDuration.newDuration = duration;
+                            }
+                        }
+                        if (laserPrefab)
+                        {
+                            laserEffect = Object.Instantiate(laserPrefab, laserMuzzle.position, laserMuzzle.rotation);
+                            laserEffect.transform.parent = laserMuzzle;
+                            laserLineComponent = laserEffect.GetComponent<LineRenderer>();
+                        }
+                    }
+                }
+            }
+            if (characterBody)
+            {
+                characterBody.SetAimTimer(duration);
+            }
+            flashTimer = 0f;
+            laserOn = true;
+        }
+
+        public override void OnExit()
+        {
+            AkSoundEngine.StopPlayingID(chargePlayID);
+            base.OnExit();
+            if (chargeEffect)
+            {
+                Destroy(chargeEffect);
+            }
+            if (laserEffect)
+            {
+                Destroy(laserEffect);
+            }
+        }
+
+        public override void Update()
+        {
+            base.Update();
+            if (laserEffect && laserLineComponent)
+            {
+                var aimRay = GetAimRay();
+                var position = laserEffect.transform.parent.position;
+                var endPoint = aimRay.GetPoint(1000f);
+                laserDirection = endPoint - position;
+                if (Physics.Raycast(aimRay, out RaycastHit raycastHit, 1000f, LayerIndex.world.mask | LayerIndex.entityPrecise.mask))
+                {
+                    endPoint = raycastHit.point;
+                }
+                laserLineComponent.SetPosition(0, position);
+                laserLineComponent.SetPosition(1, endPoint);
+                float totalWidth;
+                if (duration - age > 0.5f)
+                {
+                    totalWidth = age / duration;
+                }
+                else
+                {
+                    flashTimer -= Time.deltaTime;
+                    if (flashTimer <= 0f)
+                    {
+                        laserOn = !laserOn;
+                        flashTimer = 0.033333335f;
+                    }
+                    totalWidth = (laserOn ? 1f : 0f);
+                }
+                totalWidth *= laserMaxWidth;
+                laserLineComponent.startWidth = totalWidth;
+                laserLineComponent.endWidth = totalWidth;
+            }
+        }
+
+        public override void FixedUpdate()
+        {
+            base.FixedUpdate();
+            if (fixedAge >= duration && isAuthority)
+            {
+                FireLaserSpam fireLaserSpam = new()
+                {
+                    laserDirection = laserDirection
+                };
+                outer.SetNextState(fireLaserSpam);
+                return;
+            }
+        }
+
+        public override InterruptPriority GetMinimumInterruptPriority()
+        {
+            return InterruptPriority.Skill;
+        }
+    }
+
+    public class FireLaserSpam : BaseState
+    {
+        public static GameObject effectPrefab = Utils.Paths.GameObject.MuzzleflashGolem.Load<GameObject>();
+
+        public static GameObject hitEffectPrefab = Utils.Paths.GameObject.ExplosionGolem.Load<GameObject>();
+
+        public static GameObject tracerEffectPrefab = Utils.Paths.GameObject.TracerGolem.Load<GameObject>();
+
+        public static float damageCoefficient = 1f;
+
+        public static float blastRadius = 3f;
+
+        public static float force = 1500f;
+
+        public static float minSpread = 0f;
+
+        public static float maxSpread = 1f;
+
+        public static float maxBaseDuration = 5f;
+        public static float baseDuration = 0.5f;
+
+        public static string attackSoundString = "Play_golem_laser_fire";
+
+        public Vector3 laserDirection;
+
+        public float duration;
+
+        public Ray modifiedAimRay;
+        public float timer;
+
+        public override void OnEnter()
+        {
+            base.OnEnter();
+            duration = baseDuration / attackSpeedStat;
+            PlayCrossfade("Gesture, Additive", "FireLaserLoop", 0.25f);
+            if (characterBody)
+            {
+                characterBody.SetAimTimer(maxBaseDuration);
+            }
+        }
+
+        public override void OnExit()
+        {
+            base.OnExit();
+            PlayCrossfade("Gesture, Additive", "FireLaserEnd", 0.25f);
+        }
+
+        public override void FixedUpdate()
+        {
+            base.FixedUpdate();
+            timer += Time.fixedDeltaTime;
+            if (timer >= duration)
+            {
+                FireLaser();
+                timer = 0f;
+            }
+            if (fixedAge >= maxBaseDuration && isAuthority)
+            {
+                outer.SetNextStateToMain();
+                return;
+            }
+        }
+
+        public void FireLaser()
+        {
+            modifiedAimRay = GetAimRay();
+            modifiedAimRay.direction = laserDirection;
+            GetModelAnimator();
+            Util.PlaySound(attackSoundString, gameObject);
+            var laserMuzzle = "MuzzleLaser";
+            PlayAnimation("Gesture", "FireLaser", "FireLaser.playbackRate", duration);
+            EffectManager.SimpleMuzzleFlash(effectPrefab, gameObject, laserMuzzle, false);
+            var modelTransform = GetModelTransform();
+            if (isAuthority)
+            {
+                float num = 1000f;
+                Vector3 vector = modifiedAimRay.origin + modifiedAimRay.direction * num;
+                if (Physics.Raycast(modifiedAimRay, out RaycastHit raycastHit, num, LayerIndex.world.mask | LayerIndex.defaultLayer.mask | LayerIndex.entityPrecise.mask))
+                {
+                    vector = raycastHit.point;
+                }
+                new BlastAttack
+                {
+                    attacker = gameObject,
+                    inflictor = gameObject,
+                    teamIndex = TeamComponent.GetObjectTeam(gameObject),
+                    baseDamage = damageStat * damageCoefficient,
+                    baseForce = force * 0.2f,
+                    position = vector,
+                    radius = blastRadius,
+                    falloffModel = BlastAttack.FalloffModel.SweetSpot,
+                    bonusForce = force * modifiedAimRay.direction
+                }.Fire();
+                if (modelTransform)
+                {
+                    var childLocator = modelTransform.GetComponent<ChildLocator>();
+                    if (childLocator)
+                    {
+                        var laserMuzzleIndex = childLocator.FindChildIndex(laserMuzzle);
+                        EffectData effectData = new()
+                        {
+                            origin = vector,
+                            start = modifiedAimRay.origin
+                        };
+                        effectData.SetChildLocatorTransformReference(gameObject, laserMuzzleIndex);
+                        EffectManager.SpawnEffect(tracerEffectPrefab, effectData, true);
+                        EffectManager.SpawnEffect(hitEffectPrefab, effectData, true);
+                    }
+                }
+            }
+        }
+
+        public override InterruptPriority GetMinimumInterruptPriority()
+        {
+            return InterruptPriority.Skill;
         }
     }
 }