+++ Content/Items/Lunars/ShapedGlass.cs
@@ -8,8 +8,8 @@ namespace WellRoundedBalance.Items.Lunars
         public override string Name => ":: Items ::::: Lunars :: Shaped Glass";
         public override ItemDef InternalPickup => RoR2Content.Items.LunarDagger;
 
-        public override string PickupText => "Increase your damage by 100%... <color=#FF7F7F>BUT reduce your health by 50%.</color>";
-        public override string DescText => "Increase base damage by <style=cIsDamage>100%</style> <style=cStack>(+100% per stack)</style>. <style=cIsHealing>Reduce maximum health by 50%</style> <style=cStack>(+50% per stack)</style>.";
+        public override string PickupText => "Increase your damage by 75%... <color=#FF7F7F>BUT reduce your health by 50%.</color>";
+        public override string DescText => "Increase base damage by <style=cIsDamage>75%</style> <style=cStack>(+75% per stack)</style>. <style=cIsHealing>Reduce maximum health by 50%</style> <style=cStack>(+50% per stack)</style>.";
 
         public override void Init()
         {
@@ -29,7 +29,7 @@ namespace WellRoundedBalance.Items.Lunars
                 var stack = sender.inventory.GetItemCount(RoR2Content.Items.LunarDagger);
                 if (stack > 0)
                 {
-                    args.damageMultAdd += 1f * stack;
+                    args.damageMultAdd += 0.75f * stack;
                     args.baseCurseAdd += Mathf.Pow(1 / 0.5f, stack) - 1;
                 }
             }